refactor date-editor.js and add inclinometer statefulness
